feat(Name): 이름의 입력에 대한 검증 로직을 Name 클래스로 이동
기존에 Car 도메인에서 이름 입력에 대해 수행한 검증을 Name 도메인에서 수행하도록 수정한다.
